separated rook.cpp into a header and the corresponding file.
